Organize components and CSS files
