+++ src/CSS/Movie.css
@@ -1,15 +1,15 @@
 .movie-poster {
+  align-items: center;
   display: flex;
   flex-direction: column;
-  align-items: center;
   margin: 20px 40px;
 }
 
 .poster-img {
   border: 1.5px solid #FFDAC6;
   box-shadow: 6px 8px 10px #BABD8D;
-  width: 200px;
   height: 250px;
+  width: 200px;
 }
 
 .poster-img:hover {
@@ -23,8 +23,8 @@
 }
 
 .poster-title {
-  margin-top: 20px;
   font-size: 1.5vw;
+  margin-top: 20px;
 }
 
 .poster-rating {
